I init stack lib trackEventWithStack
