+++ lib/const.js
@@ -1,3 +1,5 @@
 // ABI for the SetupNewToken event
 export const eventSignature =
   "0x1b944478023872bf91b25a13fdba3a686fdb1bf4dbb872f850240fad4b8cc068";
+export const STACK_API_KEY = process.env.STACK_API_KEY;
+export const STACK_API_URL = "https://track.stack.so/event";
