+++ lib/indexEventsForNetwork.js
@@ -2,47 +2,15 @@ import getBlockNumber from "./getBlockNumber";
 import getLogs from "./getLogs";
 import fetch from "node-fetch";
 import dotenv from "dotenv";
+import trackEventWithStack from "./stack/trackEventWithStack";
 
 dotenv.config();
-
-const STACK_API_KEY = process.env.STACK_API_KEY;
-const STACK_API_URL = "https://track.stack.so/event";
-
-async function trackEventWithStack(network, event) {
-  try {
-    const response = await fetch(STACK_API_URL, {
-      method: "POST",
-      headers: {
-        "Content-Type": "application/json",
-        "x-api-key": STACK_API_KEY,
-      },
-      body: JSON.stringify([
-        {
-          name: "SetupNewToken",
-          account: event.address, // Assuming the contract address is relevant
-          uniqueId: `${network}-${event.transactionHash}-${event.logIndex}`,
-          pointSystemId: process.env.STACK_SYSTEM_ID, // Adjust as needed
-          points: 1, // Adjust as needed
-        },
-      ]),
-    });
-
-    if (!response.ok) {
-      throw new Error(`HTTP error! status: ${response.status}`);
-    }
-
-    console.log("Event tracked with Stack.so successfully");
-  } catch (error) {
-    console.error("Error tracking event with Stack.so:", error);
-  }
-}
 
 async function indexEventsForNetwork(network) {
   console.log(`Starting indexer for Base ${network}...`);
 
   try {
     const latestBlock = (await getBlockNumber(network)) - 10000n;
-
     console.log(`${network} - Setting up watcher for new events...`);
 
     async function checkNewEvents() {
